requirements.txt
aiogram[fast]
asyncpg
python-dotenv
